# api/openapi.yaml
openapi: 3.1.0
info:
  title: GaiaOS Canonical Carrier API
  version: 1.6.0
  description: >-
    Source-backed GaiaOS carrier with a compact primary front door, read-only
    Gaia Council/operator routing, deterministic dispatch, BrainOS support,
    DictionaryOS/YggdrasilOS context navigation, HOT/WARM/COLD, continuity,
    and diagnostic surfaces. Routing, navigation, and presentation do not grant
    domain effect authority or settle irreversible identity claims.
servers:
  - url: https://gaiaos-loader-api.onrender.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Optional when GAIAOS_API_KEY is not configured; bearer token required when enabled.
  schemas:
    GaiaAssistRequest:
      type: object
      additionalProperties: false
      required: [request]
      properties:
        request:
          type: string
          minLength: 1
          maxLength: 20000
          description: Naomi's ordinary natural-language request. This is the preferred front door.
        requested_members:
          type: array
          maxItems: 12
          default: []
          items:
            type: string
          description: Optional explicit current Gaia Council member requests.
        max_members:
          type: integer
          minimum: 1
          maximum: 32
          default: 3
        include_context:
          type: boolean
          default: true
        context_limit:
          type: integer
          minimum: 1
          maximum: 10
          default: 6
        context_depth:
          type: integer
          minimum: 0
          maximum: 2
          default: 1
    DispatchRequest:
      type: object
      additionalProperties: false
      required: [signals]
      properties:
        signals:
          type: array
          minItems: 1
          maxItems: 32
          items:
            type: string
        requested_members:
          type: array
          maxItems: 12
          default: []
          items:
            type: string
        max_members:
          type: integer
          minimum: 1
          maximum: 32
          default: 3
paths:
  /gaiaos/assist:
    post:
      operationId: assistGaiaOS
      summary: Primary GaiaOS front door
      description: >-
        Preferred ordinary-use surface. Submit Naomi's natural-language request
        once. GaiaOS conservatively resolves source-pinned semantic context,
        infers only current Gaia-native dispatch signals, preserves explicit
        member requests, and returns one compact read-only support packet. Hosts
        should prefer this over manually chaining load, context, BrainOS, and
        dispatch calls unless deep diagnostics are required.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GaiaAssistRequest'
      responses:
        '200':
          description: Compact GaiaOS front-door support packet.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Missing or invalid API key when authentication is enabled.
        '422':
          description: Invalid request.
  /gaiaos/selftest:
    get:
      operationId: selfTestGaiaOS
      summary: Self-test deployed GaiaOS organs
      description: >-
        Validate the deployed checkout, Naomi authority, council/profile
        alignment, deterministic dispatch, DictionaryOS candidates, YggdrasilOS
        graph edges, and deployed-checkout context binding. This is diagnostic
        only and does not prove external network reachability or provider effects.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: GaiaOS deployed-checkout diagnostic packet.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Missing or invalid API key when authentication is enabled.
  /gaiaos/load:
    get:
      operationId: loadGaiaOS
      summary: Load the deep GaiaOS source bundle
      description: >-
        Return the canonical bootstrap plus current Gaia Council and cognitive
        support source bundle. This is a deep diagnostic/bootstrap surface and
        is intentionally much larger than the primary /gaiaos/assist response.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verified GaiaOS bootstrap bundle.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Missing or invalid API key.
        '502':
          description: Canonical source unavailable or invalid.
  /gaiaos/council:
    get:
      operationId: getGaiaCouncil
      summary: Get the current Gaia Council
      description: >-
        Return the deep source-backed council contract, current FairyOS roster,
        operator profiles, prosody basins, dispatch matrix, EmojiOS surface,
        BrainOS/ConvoOS/ChatOS coordinates, and council command contract. The
        current Gaia-native roster is accepted for current use by Naomi and
        remains renamable.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current Gaia Council packet.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Missing or invalid API key.
        '502':
          description: Canonical source unavailable or invalid.
  /gaiaos/operator/{member}:
    get:
      operationId: getGaiaOperator
      summary: Get one current GaiaOS operator
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: member
          required: true
          schema:
            type: string
          description: Current GaiaOS operator name, case-insensitive.
      responses:
        '200':
          description: Source-backed operator profile and dispatch surface.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: Requested operator is not in the current GaiaOS roster.
        '401':
          description: Missing or invalid API key.
  /gaiaos/dispatch:
    post:
      operationId: dispatchGaiaCouncil
      summary: Route typed signals through the Gaia Council
      description: >-
        Deterministically route explicit typed signals through the current
        GaiaOS FairyOS dispatch matrix. Explicit current-member requests are
        preserved, unknowns stay visible, and the result is read-only routing,
        not domain execution.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchRequest'
      responses:
        '200':
          description: Deterministic Gaia Council dispatch packet.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Missing or invalid API key.
        '422':
          description: Invalid dispatch request.
        '502':
          description: Canonical source unavailable or invalid.
  /gaiaos/brain:
    get:
      operationId: getGaiaBrain
      summary: Get GaiaOS BrainOS support state
      description: >-
        Return the deep read-only BrainOS support fabric, Context Compass
        contract, HOT/WARM/COLD controls, response and cast modes, ConvoOS
        re-entry coordinates, and MemberContinuityOS warm-candidate surface.
        This endpoint does not create memory or execute domain effects.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current GaiaOS BrainOS and continuity support packet.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Missing or invalid API key.
        '502':
          description: Canonical source unavailable or invalid.
  /gaiaos/context:
    get:
      operationId: getGaiaContext
      summary: Resolve a natural subject through GaiaOS semantic navigation
      description: >-
        Resolve one natural-language subject against the deployed DictionaryOS
        term registry, traverse explicit YggdrasilOS graph edges, and return a
        bounded context packet. A term hit is not authority and a graph edge is
        not an effect.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: subject
          required: true
          schema:
            type: string
            minLength: 1
          description: Natural GaiaOS subject or phrase to resolve.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 10
        - in: query
          name: depth
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 2
            default: 1
      responses:
        '200':
          description: Deployed-checkout DictionaryOS/YggdrasilOS context packet.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          description: Missing or invalid API key.
        '422':
          description: Invalid or empty subject.
  /health:
    get:
      operationId: health
      summary: Health check
      responses:
        '200':
          description: Service health and source-surface availability.